+++ CHANGELOG.rst
@@ -3,6 +3,8 @@
 - Text colors other than black and white are no longer ignored
 - Textboxes have been implemented. We no longer lose the content inside of
   them.
+- Markup compatibility has been implemented. We always use the Fallback for
+  AlternateContent tags.
 
 **0.9.6**